CloudComponents/WebsocketAdapter/requirements.txt
websockets==11.0.3
aiohttp==3.9.5
python-socketio[asyncio_client]